# api/openapi.yaml
openapi: 3.1.0
info:
  title: Webhook Gateway API
  version: "1.0.0"
  description: |
    The webhook gateway's HTTP API — the full surface behind the CLI and the
    dashboard SPA. This spec is handwritten
    and is the source of truth for the Phase 7 docs-site API reference.

    ## Authentication

    Every `/api/*` endpoint takes a bearer token in `Authorization: Bearer <token>`.
    Two credential types are accepted:

    - **Admin password** — the `ADMIN_PASSWORD` the gateway boots with. Passes
      every scope check; the only credential that may manage API keys.
    - **API key** (`whg_...`) — scoped. Each key carries a subset of
      `read`, `write`, `replay`, `tunnel`. A request whose endpoint needs a
      scope the key lacks gets `403`; an unknown or revoked key gets `401`.

    The dashboard SPA instead uses a **session cookie** (`whg_session`) obtained
    from `POST /api/auth/login`, so no credential is ever held in JavaScript.
    A session is the admin password already proven, so it is accepted on every
    endpoint the admin password is, ncluding API-key management.

    The `/ingest/{path}` endpoint is unauthenticated by bearer token — it is
    authenticated by the source path plus the provider signature.

    ## Errors

    All error responses share one envelope: `{"error": "<message>"}`, including
    those emitted by the auth middleware.
servers:
  - url: http://localhost:8080
    description: Local / self-hosted gateway
...
paths:
...
  /api/api-keys:
    get:
      tags: [api-keys]
      summary: List API keys
      description: Admin-password only. Never returns key plaintext or hash; revoked keys remain listed as an audit trail.
      security: [{ adminAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ApiKey" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [api-keys]
      summary: Create an API key
      description: Admin-password only. The plaintext key is returned exactly once, here.
      security: [{ adminAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, scopes]
              properties:
                name: { type: string }
                scopes:
                  type: array
                  items: { type: string, enum: [read, write, replay, tunnel] }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiKey"
                  - type: object
                    properties:
                      key: { type: string, description: Plaintext key, shown once. }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/api-keys/{id}:
    delete:
      tags: [api-keys]
      summary: Revoke an API key
      description: Admin-password only. Soft-delete; the row is kept with revoked_at set.
      security: [{ adminAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "204": { description: Revoked }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/auth/login:
    post:
      tags: [auth]
      summary: Start a dashboard session
      description: |
        Exchanges the admin password for an `whg_session` cookie (HttpOnly,
        SameSite=Lax, 7-day sliding expiry). Intended for the dashboard SPA;
        programmatic clients should use an API key instead.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password: { type: string }
      responses:
        "200":
          description: Session cookie set
          headers:
            Set-Cookie:
              schema: { type: string }
          content:
            application/json:
              schema:
                type: object
                properties:
                  authenticated: { type: boolean }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /api/auth/session:
    get:
      tags: [auth]
      summary: Check the current session
      description: |
        Returns 200 when the request carries a valid session cookie, 401
        otherwise. The SPA calls this on load to choose between the app and the
        login page. A cookie past half its life is renewed on the response.
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: Session is valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  authenticated: { type: boolean }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: An API key (whg_...) or the admin password. Scopes listed per-operation apply to API keys.
    adminAuth:
      type: http
      scheme: bearer
      description: The admin password only. API keys are rejected on these endpoints.
    sessionCookie:
      type: apiKey
      in: cookie
      name: whg_session
      description: |
        The dashboard's login cookie, obtained from POST /api/auth/login. It
        carries admin authority, so it is accepted anywhere the admin password
        is — including the admin-only API-key endpoints.
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    Error:
      description: Request error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Unauthorized:
      description: Missing, unknown, or revoked credential
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  schemas:
    ErrorEnvelope:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    ApiKey:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        key_prefix: { type: string, description: First 12 chars, for display. }
        scopes:
          type: array
          items: { type: string, enum: [read, write, replay, tunnel] }
        created_at: { type: string, format: date-time }
        last_used_at: { type: [string, "null"], format: date-time }
        revoked_at: { type: [string, "null"], format: date-time }